+++ src/action_machine/Core/components/tools_box_factory.py
@@ -119,7 +119,6 @@ class ToolsBoxFactory:
             run_child=run_child,
             factory=factory,
             resources=resources,
-            context=context,
             log=log,
             nested_level=nest_level,
             rollup=rollup,